(maint) added logs for get_expected_value method
Prior to this PR, if get_expected_value method returns False, the test that uses this method simply fails without any debug logs, hence it is very hard to trace why it fails.

This PR addes debugs logs printing the output of `openstack show` command so it will be easier to see what the return values are.