ASC-705 Call list on filter objects for python 3
This commit calls `list` on the filter objects used in the `parse_table`
helper. In python 3.x `filter` constructs an iterator. Calling `len` on
these object raised an exception. Converting them to a list of values
derived from the filter allows the unit tests for this helper to pass
when run with python 3.x.