ASC-765 Change openstack_name_list to return json
This commit changes the `openstack_name_list` method to return the json
list payload from the `openstack <resource-type> list` command rather
than the string payload previously returned. The intention of this is to
make the data easier to validate against rather than just being able to
test for sub-string inclusion.

The name of the method has also been changed to
`get_resource_list_by_name` to bring its naming convention in closer
alignment with the rest of the helpers.