MAINT - Fix data gather get_resource_list_by_name
This commit fixes a bug in the `get_resource_list_by_name` helper.

The helper proclaims that it will gather dictionary data for the given
resource type. However, the underlying call to the `openstack` cli does
not specify that JSON data should be returned. So, the helper will
always return an empty list even when there is valid data.

This bug is corrected and a unit test has been added to prevent this
error from regressing in the future.