ASC-1559 Fix Version File Lookup
Turns out that the CI scripts will lay down a copy of '/etc/openstack-release'
if it doesn't exist somewhere in the post process. However, it already exists
then it assumes that the file is readable. Adding the '[default]' INI section
to the file makes it fail to import which in turn causes the post script to
fail with the result of no qTest results being published.

So, the modified file is now being appended with the extension of '.ini' to
differentiate it from the original file.