ASC-1627 Add get_cinder_major_version helper
This commit adds a helper method called `get_cinder_major_version` that
returns the major version of cinder as inspected on the utility
container of an RPC-O installation.  Supporting tests are included as
well.

This helper makes use of the [packaging](https://packaging.pypa.io/en/latest/)
python module.  So, the requirements and constraints files are updated
with this dependency.